docs(real_data_checks): add OBS→NONE reprojection sweep tools and parity closure analysis
New scripts:
- rank_obsnone_reproject_scan.py  — closeness ranking across reproject_* outputs
- plot_obsnone_relative_error_panel.py — vector-magnitude relative error panel
- plot_obsnone_component_metric_panels.py — per-component abs-residual and sym-diff panels
- export_none_base_maps_to_fits.py — export NONE base maps to IDL-readable FITS MEF

Updated:
- check_idl2py_stage_parity.py — add --reproject-algorithm flag; project-specific artifact roots
- pyampp/gxbox/gx_fov2box.py — --reproject-algorithm CLI param; legacy NLFFF lib ctypes patch
- pyampp/io/_sav_convert.py — promote corona/base fields to float64; line indices to int32
- pyampp/io/model.py — model load and metadata path fixes
- real_data_checks/README.md — document new tools, reprojection sweep results, closure analysis

Analysis findings documented in README:
- adaptive > exact > interpolation for OBS→NONE base-map fidelity
- boundary injection test confirms downstream (POT→CHR) pipeline is correct
- OBS→NONE gap isolated as dedicated post-merge investigation target

+++ pyampp/io/_sav_convert.py
@@ -160,15 +160,15 @@ def _to_line_flat(arr: Any, dtype) -> np.ndarray:
 
 def _normalize_czyx_from_components_or_bcube(box: Any) -> np.ndarray | None:
     if _has_field(box, "BCUBE"):
-        b = np.asarray(box["BCUBE"], dtype=np.float32)
+        b = np.asarray(box["BCUBE"], dtype=np.float64)
         if b.ndim == 4 and b.shape[0] == 3:
             return b
         if b.ndim == 4 and b.shape[-1] == 3:
             return np.moveaxis(b, -1, 0)
     if _has_field(box, "BX") and _has_field(box, "BY") and _has_field(box, "BZ"):
-        bx = np.asarray(box["BX"], dtype=np.float32)
-        by = np.asarray(box["BY"], dtype=np.float32)
-        bz = np.asarray(box["BZ"], dtype=np.float32)
+        bx = np.asarray(box["BX"], dtype=np.float64)
+        by = np.asarray(box["BY"], dtype=np.float64)
+        bz = np.asarray(box["BZ"], dtype=np.float64)
         if bx.ndim == 3 and by.shape == bx.shape and bz.shape == bx.shape:
             # scipy.readsav restores GX BOX component cubes in (z, y, x), which
             # already matches the canonical HDF5 3D axis order.
@@ -362,8 +362,8 @@ def build_h5_from_sav(sav_path: Path, out_h5: Path, template_h5: Path | None = N
                 _replace_dataset(g_lines, "phys_length", _to_line_flat(box["PHYSLENGTH"], np.float64))
             if _has_field(box, "STATUS"):
                 _replace_dataset(g_lines, "voxel_status", _to_line_flat(box["STATUS"], np.uint8))
-            _replace_dataset(g_lines, "start_idx", _to_line_flat(box["STARTIDX"], np.int64))
-            _replace_dataset(g_lines, "end_idx", _to_line_flat(box["ENDIDX"], np.int64))
+            _replace_dataset(g_lines, "start_idx", _to_line_flat(box["STARTIDX"], np.int32))
+            _replace_dataset(g_lines, "end_idx", _to_line_flat(box["ENDIDX"], np.int32))
             _replace_dataset(g_lines, "dr", dr.astype(np.float64))
 
         if has_chromo:
@@ -434,7 +434,7 @@ def build_h5_from_sav(sav_path: Path, out_h5: Path, template_h5: Path | None = N
                 "corona_base": int(_field(box, "CORONA_BASE", 0)),
             }
             if _has_field(box, "STARTIDX"):
-                id_input["start_idx"] = _to_line_flat(box["STARTIDX"], np.int64)
+                id_input["start_idx"] = _to_line_flat(box["STARTIDX"], np.int32)
             if _has_field(box, "CHROMO_IDX"):
                 id_input["chromo_idx"] = np.asarray(box["CHROMO_IDX"], dtype=np.int64)
             if _has_field(box, "CHROMO_T"):
